refactor: modalv2에서 modal로 이름변경
